fix: jwt utilの統合とsecretの呼び出し
