--- internal/utils/jwt.go
@@ -1,42 +1,0 @@
-package utils
-
-import (
-	"github.com/dgrijalva/jwt-go"
-)
-
-var jwtKey = []byte("your_secret_key")
-
-type Claims struct {
-	UID string `json:"uid"`
-	jwt.StandardClaims
-}
-
-func GenerateJWT(uid string) (string, error) {
-	// expirationTime := time.Now().Add(24 * time.Hour)
-	claims := &Claims{
-		UID:            uid,
-		StandardClaims: jwt.StandardClaims{
-			// ExpiresAt: expirationTime.Unix(),
-		},
-	}
-
-	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
-	return token.SignedString(jwtKey)
-}
-
-func ParseJWT(tokenStr string) (*Claims, error) {
-	claims := &Claims{}
-	token, err := jwt.ParseWithClaims(tokenStr, claims, func(token *jwt.Token) (interface{}, error) {
-		return jwtKey, nil
-	})
-
-	if err != nil {
-		return nil, err
-	}
-
-	if !token.Valid {
-		return nil, jwt.ErrSignatureInvalid
-	}
-
-	return claims, nil
-}
